fix equations and bugs in graph
Fixed some bugs with the bar graph display. The graphs are colored from red to green, with red indicating the lowest expected SOC increase and green with the highest expected increase. The bar graphs display the rounded values of the expected SOC increases, and the labels under each bar graph are shortened to be more readable.

Fixed the equation used to calculate SOC energy increase to make the output values in percent more realistic.