fixed time issue and changed time segment appearance on graph
Fixed the time zone conversion issue from UTC to central. The time zone can be changed if the targeted area for irradiance is changed. Changed the x-axis time segment labels to include the start and end time to avoid confusion with entering different time lengths for time segments, since total SOC increase may be higher for one segment, but the average is lower. This is made more clear with changing the labels.